Added code for fault dictionary
The algorithm maintains a fault dictionary and tests the memristor for every possible IO switch vector. If the current is equivalent to expected current, the bitline and wordline are eliminated to be possible faulty and dictionary stores 0 (not faulty) for the whole bitline and wordline. If the current value differs from the expected, then dictionary stores 1 for that memristor (faulty) as this may or may not be the faulty memristor.
This is a brute force algorithm that only works for finding one faulty memristor.